dev: item; redid pools; more sfx; prep bubble images

+++ Assets/Game/Scripts/STG/Bullet/PlayerBullet.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class PlayerBullet : Bullet
 {
-    public int damage = 1;
+    public ushort damage = 1;
 
     public override void CrossCameraBoundary(Collider2D other)
     {
